Final Product

+++ PrivateLMS/Models/BookRating.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int UserId { get; set; }
         public int BookId { get; set; }
-        public float Rating { get; set; } // 1 to 5
+        public int Rating { get; set; } // 1 to 5
         public string? Review { get; set; } 
         public DateTime RatedOn { get; set; }
